Move name binding to its own data structure
This avoids unnecessary clones
