+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod compiler;
 pub mod errors;
 pub mod parser;
+pub mod resolver;
 #[cfg(test)]
 mod test;
