+++ src/test.rs
@@ -1,5 +1,7 @@
+use crate::resolver::Resolver;
 use crate::{compiler::Compiler, parser::Parser};
 use std::path::Path;
+use std::process::exit;
 
 fn evaluate_example(fname: &Path) -> String {
     let mut compiler = Compiler::new();
@@ -9,11 +11,19 @@ fn evaluate_example(fname: &Path) -> String {
     compiler.add_file(&fname.to_string_lossy(), &contents);
 
     let parser = Parser::new(compiler, span_offset);
-
     compiler = parser.parse();
-    compiler.resolve();
 
-    compiler.display_state()
+    let mut result = compiler.display_state();
+
+    if !compiler.errors.is_empty() {
+        return result;
+    }
+
+    let mut resolver = Resolver::new(&compiler);
+    resolver.resolve();
+    result.push_str(&resolver.display_state());
+
+    result
 }
 
 #[test]
